[refactor] 불필요 동작 제거 및 패키지 이동
