--- movie/infra/src/main/kotlin/com/example/movie/persistence/movie/projection/CacheMixin.kt
@@ -1,1 +1,0 @@
-package com.example.movie.persistence.movie.projection
